Update compulsory validation(form).html
Requiring an Input
Sometimes we have fields in our <form>s which are not optional, i.e. there must be information provided before we can submit it. To enforce this rule, we can add the required attribute to an <input> element.

This renders a text box, and if we try to submit the <form> without filling it out we get this message:
Do you have any dietary restrictions?

Please fill out this box
The styling of the message varies from browser to browser; the picture above depicts the message in a Chrome browser. We’ll also continue to show these messages as they appear in Chrome in later exercises.