Create Checking Text Length(form).html
Checking Text Length
In the previous exercise, we were able to use min and max to set acceptable minimum and maximum values in a number field. But what about text fields? There are certainly cases where we wouldn’t want our users typing more than a certain number of characters (think about the character cap for messages on Twitter). We might even want to set a minimum number of characters. Conveniently, there are built-in HTML5 validations for these situations.
To set a minimum number of characters for a text field, we add the minlength attribute and a value to set a minimum value. Similarly, to set the maximum number of characters for a text field, we use the maxlength attribute and set a maximum value. Let’s take a look at these attributes in code:

If a user tries to submit the <form> with less than the set minimum, this message appears: below required minimum character. And if a user tries to type in more than the maximum allowed number of characters, they don’t get a warning message, but they can’t type it in!